REFACTOR: vscode - move function to new file
