+++ src/refactorings/simplifyConditional.ts
@@ -1,6 +1,6 @@
-import * as tsutils from 'tsutils';
 import * as ts from 'typescript/lib/tsserverlibrary';
 import { Logger } from '../logger';
+import { tryGetTargetExpression } from './tryGetTargetExpression';
 
 export const name = 'Simplify Conditional';
 export const simplifyConstantBooleanExpression = 'simplify_constant_boolean_expression';
@@ -95,24 +95,7 @@ function simplifyBinaryExpression(expression: ts.BinaryExpression): ts.Expressio
   return ts.updateBinary(expression, simplifiedLeft, simplifiedRight);
 }
 
-function tryGetTargetExpression(
-  logger: Logger,
-  sourceFile: ts.SourceFile,
-  positionOrRange: number | ts.TextRange
-): ts.BinaryExpression | null {
-  const startPos = typeof positionOrRange === 'number' ? positionOrRange : positionOrRange.pos;
-
-  const token = tsutils.getTokenAtPosition(sourceFile, startPos);
-
-  if (token === undefined || token.parent === undefined) {
-    logger.error(`No token at given position ${startPos}`);
-    return null;
-  }
-
-  return findLargestPossibleBinaryExpressionParentNode(token);
-}
-
-function findLargestPossibleBinaryExpressionParentNode(
+export function findLargestPossibleBinaryExpressionParentNode(
   startNode: ts.Node
 ): ts.BinaryExpression | null {
   const firstBinaryExpression = getNextParentBinaryExpression(startNode);
